requirements.txt
Django==2.0
django-ckeditor==5.3.1
django-crispy-forms==1.7.0
django-js-asset==0.1.1
django-pygments==0.1
et-xmlfile==1.0.1
jdcal==1.3
lml==0.0.1
mistune==0.8.3
mysqlclient==1.3.12
olefile==0.44
openpyxl==2.5.0
Pillow==4.3.0
pyexcel==0.5.7
pyexcel-io==0.5.6
pyexcel-webio==0.1.4
Pygments==2.2.0
pytz==2017.3
texttable==1.2.1
